pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "trustfall-poc"
version = "0.1.0"
description = "TRUSTFALL benchmark POC — agent safety across composed enterprise systems"
authors = [{name = "Shiva Reddy Peddireddy"}]
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.32",
  "pydantic>=2.9",
  "sqlalchemy>=2.0",
  "httpx>=0.27",
  "openai>=1.50",
  "anthropic>=0.40",
  "rich>=13.9",
  "typer>=0.13",
  "pyyaml>=6.0",
  "jinja2>=3.1",
]

[project.optional-dependencies]
dev = ["pytest>=8", "pytest-asyncio>=0.24", "ruff>=0.8"]

[tool.setuptools.packages.find]
where = ["."]
include = ["corpsim*", "harness*", "baselines*", "dashboard*"]
